chore(deps): ⬆️ Update @types/mithril to version 2.0.13
Updated the devDependency @types/mithril to version 2.0.13 in package.json and pnpm-lock.yaml. This was necessitated by the need to keep up with the latest changes and improvements in the Mithril type definitions. Ensuring the type definitions used in the project align with the current state of Mithril enhances type safety and improves developer experience.