chore(deps): Update package versions and add new devDependency
Updated the version of lockfile from '6.1' to '6.0' and the versions of '@types/mithril' and 'typescript' in both 'pnpm-lock.yaml' and 'package.json'. Additionally, 'type-fest' was added as a new devDependency. The change was necessary to ensure compatibility with the updated package versions and to use the new functions from the added 'type-fest' package. The directory of 'typings.d.ts' file was also corrected and the version of packageManager updated to 'pnpm@8.7.5'.